Скотин Александр. Задача 1. Вариант 4. Минимальное значение элементов вектора.

+++ tasks/task_1/skotin_a_min_of_vector_elements/min_of_vector_elements.cpp
@@ -23,7 +23,7 @@ int VectorMinFinder::sequentialMin(const std::vector<int>& data) {
     return *std::min_element(data.begin(), data.end());
 }
 
-std::vector<int> VectorMinFinder::generateRandomVector(size_t size, int min, int max) {
+std::vector<int> VectorMinFinder::generateRandomVector(int size, int min, int max) {
     std::vector<int> vec(size);
     std::random_device rd;
     std::mt19937 gen(rd());
