+++ tasks/task_1/skotin_a_min_of_vector_elements/min_of_vector_elements.h
@@ -12,7 +12,7 @@ class VectorMinFinder {
 
     int parallelMin(const std::vector<int>& data);
     static int sequentialMin(const std::vector<int>& data);
-    static std::vector<int> generateRandomVector(size_t size, int min = -100, int max = 100);
+    static std::vector<int> generateRandomVector(int size, int min = -100, int max = 100);
 };
 
 #endif  // TASKS_TASK_1_SKOTIN_A_MIN_OF_VECTOR_ELEMENTS_MIN_OF_VECTOR_ELEMENTS_H_
